feat: implement cjsToEsm
